requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
supabase==2.5.1
pillow==10.4.0
